refactor(sdk): extract shared Windows terminal module
The Windows console handles, virtual-terminal modes, and the input/resize event
pump lived inside the windows arm of `attach.rs`, but `ssh.rs` drove its own
interactive session by reaching into them through `attach::agent::`.

Move them into `sandbox/terminal/`, so `attach.rs` keeps only the attach loop
and both consumers import from a named module instead. Pure code movement with
no behaviour change.
