+++ sdk/rust/lib/sandbox/mod.rs
@@ -20,6 +20,8 @@ mod patch;
 mod reap;
 #[cfg(feature = "ssh")]
 pub mod ssh;
+#[cfg(windows)]
+pub(crate) mod terminal;
 mod types;
 pub(crate) mod upper;
