+++ sdk/rust/lib/sandbox/terminal/mod.rs
@@ -1,0 +1,13 @@
+//! Host terminal integration shared by interactive sandbox sessions.
+
+#[cfg(windows)]
+mod windows;
+
+//--------------------------------------------------------------------------------------------------
+// Re-Exports
+//--------------------------------------------------------------------------------------------------
+
+#[cfg(windows)]
+pub(crate) use windows::{
+    WindowsTerminalEvent, WindowsTerminalEventPump, WindowsTerminalGuard, current_terminal_size,
+};
